Live update for update, delete & widgetisation of NoteCard
